+++ core/broadleaf-framework/src/main/java/org/broadleafcommerce/core/catalog/domain/FeaturedProduct.java
@@ -17,6 +17,8 @@
 package org.broadleafcommerce.core.catalog.domain;
 
 
+import java.math.BigDecimal;
+
 public interface FeaturedProduct extends PromotableProduct {
 
     Long getId();
@@ -31,9 +33,9 @@ public interface FeaturedProduct extends PromotableProduct {
 
     void setProduct(Product product);
 
-    void setSequence(Long sequence);
+    void setSequence(BigDecimal sequence);
     
-    Long getSequence();
+    BigDecimal getSequence();
 
     String getPromotionMessage();
