+++ core/broadleaf-framework/src/main/java/org/broadleafcommerce/core/catalog/domain/RelatedProduct.java
@@ -17,6 +17,8 @@
 package org.broadleafcommerce.core.catalog.domain;
 
 
+import java.math.BigDecimal;
+
 public interface RelatedProduct extends PromotableProduct {
     
     public Long getId();
@@ -29,7 +31,7 @@ public interface RelatedProduct extends PromotableProduct {
 
     public String getPromotionMessage();
 
-    public Long getSequence();
+    public BigDecimal getSequence();
 
     public void setId(Long id);
 
@@ -41,5 +43,5 @@ public interface RelatedProduct extends PromotableProduct {
 
     public void setPromotionMessage(String promotionMessage);
 
-    public void setSequence(Long sequence);
+    public void setSequence(BigDecimal sequence);
 }
